+++ zenml/steps/split/split_step_test.py
@@ -18,10 +18,10 @@ import random
 import pytest
 import tensorflow as tf
 
-from zenml.core.steps.split import CategoricalDomainSplit
-from zenml.core.steps.split import CategoricalRatioSplit
-from zenml.core.steps.split import NoSplit
-from zenml.core.steps.split import RandomSplit
+from zenml.steps.split import CategoricalDomainSplit
+from zenml.steps.split import CategoricalRatioSplit
+from zenml.steps.split import NoSplit
+from zenml.steps.split import RandomSplit
 
 
 @pytest.fixture
