+++ zenml/steps/split/utils_test.py
@@ -16,7 +16,7 @@
 import pytest
 import tensorflow as tf
 
-from zenml.core.steps.split.utils import get_categorical_value, \
+from zenml.steps.split.utils import get_categorical_value, \
     partition_cat_list
